fix the bug with tailwind
