+++ src/components/ExpenseList.jsx
@@ -22,11 +22,11 @@ const ExpenseList = () => {
     <>
       <input
         type="text"
-        className="w-full text-[1.4rem]"
+        className="w-[70%] text-[1.4rem] rounded-lg p-[1rem] my-[1rem] bg-[#171528] border-2 border-white text-white"
         placeholder="Type to search..."
         onChange={handleChange}
       />
-      <ul className="">
+      <ul className="bg-[#D6DBD2] p-10 my-[3rem] rounded-xl w-[70%]">
         {filteredExpenses.map((expense) => (
           <ExpenseItem
             key={expense.id}
