+++ src/components/Remaining.jsx
@@ -8,10 +8,8 @@ const RemainingBudget = () => {
     return (total += item.cost);
   }, 0);
 
-  const alertType = totalExpenses > budget ? "alert-danger" : "alert-success";
-
   return (
-    <div className={`alert p-4 ${alertType}`}>
+    <div>
       <span>Remaining: £{budget - totalExpenses}</span>
     </div>
   );
