+++ src/components/ViewBudget.jsx
@@ -2,7 +2,11 @@ const ViewBudget = (props) => {
   return (
     <>
       <span>Budget: £{props.budget}</span>
-      <button type="button" className="btn" onClick={props.handleEditClick}>
+      <button
+        type="button"
+        className="text-[#E75A7C]"
+        onClick={props.handleEditClick}
+      >
         Edit
       </button>
     </>
